Added iconsax in dependency

+++ lib/models/world_stats.dart
@@ -9,7 +9,7 @@ class WorldStatsModel {
   int? active;
   int? critical;
   int? casesPerOneMillion;
-  double? deathsPerOneMillion;
+  int? deathsPerOneMillion;
   int? tests;
   double? testsPerOneMillion;
   int? population;
